initial version refactor
