+++ experiment/builder_runner.py
@@ -29,6 +29,7 @@ from google.cloud import storage
 
 from experiment import oss_fuzz_checkout, textcov
 from experiment.benchmark import Benchmark
+from experiment.fuzz_target_error import SemanticCheckResult
 from experiment.workdir import WorkDirs
 from llm_toolkit import code_fixer
 from llm_toolkit.models import DefaultModel
@@ -38,6 +39,20 @@ JCC_DIR = '/usr/local/bin'
 
 RUN_TIMEOUT: int = 30
 CLOUD_EXP_MAX_ATTEMPT = 5
+
+LIBFUZZER_MODULES_LOADED_REGEX = re.compile(
+    r'^INFO:\s+Loaded\s+\d+\s+(modules|PC tables)\s+\((\d+)\s+.*\).*')
+LIBFUZZER_COV_REGEX = re.compile(r'.*cov: (\d+) ft:')
+LIBFUZZER_CRASH_TYPE_REGEX = re.compile(r'.*Test unit written to.*')
+LIBFUZZER_COV_LINE_PREFIX = re.compile(r'^#(\d+)')
+LIBFUZZER_STACK_FRAME_LINE_PREFIX = re.compile(r'^\s+#\d+')
+CRASH_EXCLUSIONS = re.compile(r'.*(slow-unit-|timeout-|leak-|oom-).*')
+CRASH_STACK_WITH_SOURCE_INFO = re.compile(r'in.*:\d+:\d+$')
+
+LIBFUZZER_LOG_STACK_FRAME_LLVM = '/src/llvm-project/compiler-rt'
+LIBFUZZER_LOG_STACK_FRAME_CPP = '/usr/local/bin/../include/c++'
+
+EARLY_FUZZING_ROUND_THRESHOLD = 3
 
 
 @dataclasses.dataclass
@@ -52,12 +67,17 @@ class BuildResult:
 
 @dataclasses.dataclass
 class RunResult:
+  succeeded: bool = False
   coverage_summary: dict = dataclasses.field(default_factory=dict)
   coverage: Optional[textcov.Textcov] = None
   log_path: str = ''
   corpus_path: str = ''
   coverage_report_path: str = ''
   reproducer_path: str = ''
+  cov_pcs: int = 0
+  total_pcs: int = 0
+  crashes: bool = False
+  semantic_check: SemanticCheckResult = SemanticCheckResult(SemanticCheckResult.NOT_APPLICABLE)
 
   def dict(self):
     return dataclasses.asdict(self)
@@ -119,6 +139,132 @@ class BuilderRunner:
       return False
     return True
 
+  def _parse_stacks_from_libfuzzer_logs(self,
+                                        lines: list[str]) -> list[list[str]]:
+    """Parse stack traces from libFuzzer logs."""
+    # TODO (dongge): Use stack parsing from ClusterFuzz.
+    # There can have over one thread stack in a log.
+    stacks = []
+
+    # A stack -> a sequence of stack frame lines.
+    stack, stack_parsing = [], False
+    for line in lines:
+      is_stack_frame_line = LIBFUZZER_STACK_FRAME_LINE_PREFIX.match(
+          line) is not None
+      if (not stack_parsing) and is_stack_frame_line:
+        # First line.
+        stack_parsing = True
+        stack = [line.strip()]
+      elif stack_parsing and is_stack_frame_line:
+        # Middle line(s).
+        stack.append(line.strip())
+      elif stack_parsing and (not is_stack_frame_line):
+        # Last line.
+        stack_parsing = False
+        stacks.append(stack)
+
+    # Last stack.
+    if stack_parsing:
+      stacks.append(stack)
+
+    return stacks
+
+  def _parse_fuzz_cov_info_from_libfuzzer_logs(
+      self,
+      lines: list[str]) -> tuple[Optional[int], Optional[int], Optional[int]]:
+    """Parses cov of INITED & DONE, and round number from libFuzzer logs."""
+    initcov, donecov, lastround = None, None, None
+
+    for line in lines:
+      if line.startswith('#'):
+        # Parses cov line to get the round number.
+        match = LIBFUZZER_COV_LINE_PREFIX.match(line)
+        roundno = int(match.group(1)) if match else None
+
+        if roundno is not None:
+          lastround = roundno
+          if 'INITED' in line:
+            initcov = int(line.split('cov: ')[1].split(' ft:')[0])
+          elif 'DONE' in line:
+            donecov = int(line.split('cov: ')[1].split(' ft:')[0])
+
+    return initcov, donecov, lastround
+
+  def _stack_func_is_of_testing_project(self, stack_frame: str) -> bool:
+    return (bool(CRASH_STACK_WITH_SOURCE_INFO.match(stack_frame)) and
+            LIBFUZZER_LOG_STACK_FRAME_LLVM not in stack_frame and
+            LIBFUZZER_LOG_STACK_FRAME_CPP not in stack_frame)
+
+  def _parse_libfuzzer_logs(self, log_handle) -> tuple[int, int, bool, SemanticCheckResult]:
+    """Parses libFuzzer logs."""
+    lines = None
+    try:
+      fuzzlog = log_handle.read(-1)
+      # Some crashes can mess up the libfuzzer output and raise decode error.
+      fuzzlog = fuzzlog.decode('utf-8', errors='ignore')
+      lines = fuzzlog.split('\n')
+    except MemoryError as e:
+      # Some logs from abnormal fuzz targets are too large to be parsed.
+      logging.error('%s is too large to parse: %s', log_handle.name, e)
+      return 0, 0, False, SemanticCheckResult(SemanticCheckResult.LOG_MESS_UP)
+
+    cov_pcs, total_pcs, crashes = 0, 0, False
+
+    for line in lines:
+      m = LIBFUZZER_MODULES_LOADED_REGEX.match(line)
+      if m:
+        total_pcs = int(m.group(2))
+        continue
+
+      m = LIBFUZZER_COV_REGEX.match(line)
+      if m:
+        cov_pcs = int(m.group(1))
+        continue
+
+      m = LIBFUZZER_CRASH_TYPE_REGEX.match(line)
+      if m and not CRASH_EXCLUSIONS.match(line):
+        crashes = True
+        continue
+
+    initcov, donecov, lastround = self._parse_fuzz_cov_info_from_libfuzzer_logs(
+        lines)
+
+    # NOTE: Crashes from incorrect fuzz targets will not be counted finally.
+
+    if crashes:
+      symptom = SemanticCheckResult.extract_symptom(fuzzlog)
+      crash_stacks = self._parse_stacks_from_libfuzzer_logs(lines)
+
+      # FP case 1: fuzz target crashes at init or first few rounds.
+      if lastround is None or lastround <= EARLY_FUZZING_ROUND_THRESHOLD:
+        # No cov line has been identified or only INITED round has been passed.
+        # This is very likely the false positive cases.
+        return cov_pcs, total_pcs, True, \
+               SemanticCheckResult(SemanticCheckResult.FP_NEAR_INIT_CRASH,\
+                             symptom, crash_stacks)
+
+      # FP case 2: 1st func of the 1st thread stack is in fuzz target.
+      if len(crash_stacks) > 0:
+        first_stack = crash_stacks[0]
+        # Check the first stack frame of the first stack only.
+        for stack_frame in first_stack[:1]:
+          if self._stack_func_is_of_testing_project(stack_frame):
+            if 'LLVMFuzzerTestOneInput' in stack_frame:
+              return cov_pcs, total_pcs, True, \
+                     SemanticCheckResult(SemanticCheckResult.FP_TARGET_CRASH,\
+                                   symptom, crash_stacks)
+            break
+
+    else:
+      # Another error fuzz target case: no cov increase.
+      if initcov is not None and donecov is not None:
+        if initcov == donecov:
+          return cov_pcs, total_pcs, True, SemanticCheckResult(
+              SemanticCheckResult.NO_COV_INCREASE)
+
+    return cov_pcs, total_pcs, crashes, SemanticCheckResult(
+        SemanticCheckResult.NO_SEMANTIC_ERR)
+
   def build_and_run(self, generated_project: str, target_path: str,
                     iteration: int) -> tuple[BuildResult, Optional[RunResult]]:
     """Builds and runs the fuzz target for fuzzing."""
@@ -144,6 +290,12 @@ class BuilderRunner:
         self.work_dirs.run_logs_target(benchmark_target_name, iteration))
     run_result.coverage, run_result.coverage_summary = (self.get_coverage_local(
         generated_project, benchmark_target_name))
+
+    # Parse libfuzzer logs to get fuzz target runtime details.
+    with open(self.work_dirs.run_logs_target(benchmark_target_name, iteration), 'rb') as f:
+      run_result.cov_pcs, run_result.total_pcs, run_result.crashes, run_result.semantic_check = self._parse_libfuzzer_logs(f)
+      run_result.succeeded = not run_result.semantic_check.has_err
+
     return build_result, run_result
 
   def run_target_local(self, generated_project: str, benchmark_target_name: str,
@@ -486,6 +638,11 @@ class CloudBuilderRunner(BuilderRunner):
                 # Don't include other functions defined in the target code.
                 re.compile(r'^' + re.escape(target_basename) + ':')
             ])
+
+    # Parse libfuzzer logs to get fuzz target runtime details.
+    with open(self.work_dirs.run_logs_target(generated_target_name, iteration), 'rb') as f:
+      run_result.cov_pcs, run_result.total_pcs, run_result.crashes, run_result.semantic_check = self._parse_libfuzzer_logs(f)
+      run_result.succeeded = not run_result.semantic_check.has_err
 
     return build_result, run_result
